Refactor: Add exhaustive Zone parity tests to test_tx_parity.py
